add energy scopes
This bot is now also used by the teg bot to obtain authorization.
Therefore, we ask for energy scopes as well as vehicle scopes.
Importantly, we prompt the user for missing scopes.  Without that, the
auth seems to go through OK, but then when calling APIs, the scopes are
missing and an error would result.